backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
nflreadpy==0.1.3
platformdirs==4.4.0
polars==1.33.1
psycopg2-binary==2.9.9
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
